+++ source/net/FFL_NetUtils.cpp
@@ -15,7 +15,7 @@
 
 namespace FFL {
 
-	void FFL_parseHostport(const String& url, String& host, int32_t& port){
+	void FFL_parseHostport(const String& url, String& host, int16_t& port){
 		int pos = url.find(":");
 		if (pos >=0) {			
 			host.append(url.string(), pos - 1);
